add force flag to remove action
when removing a machine the user is asked to confirm the deletion. However, in some cases, like an extension in podman desktop, the user has already been asked for a confirmation and we just need a way to delete the machine without prompts. This patch adds the force flag to achieve this.

Signed-off-by: Luca Stocchi <lstocchi@redhat.com>