return error if start non-existing machine
this patch removes the creation process if the machine that wants to be started does not exist.

The main reason is that we are going to support multiple machines in the near future and having such a functionality can be problematic.
What if the user just write the machine name wrongly? It will start an unwanted init process.

This commit just removes it and return an error message if the machine does not exist.

Signed-off-by: Luca Stocchi <lstocchi@redhat.com>