run e2e in gh workflow
it adds an e2e step to run e2e tests.
It only runs on linux for the moment. Windows would follow up but macOS does not currently support nested virtualization so we should find an alternative.

It only runs on ubuntu 24.04+ because it has a newer version of qemu (ubuntu 22.04 still uses qemu v6.2.0)  which supports the '-netdev stream'.

The setup qemu step is quite important as we should add the runner user to the kvm group. However to avoid to have to logout and login again to make the group change effective we widen permissions to the /dev/kvm device. Idea taken from https://stackoverflow.com/questions/37300811/android-studio-dev-kvm-device-permission-denied/61984745#61984745

Gvproxy gets downloaded into the tools folder and we leverage the env CONTAINERS_HELPER_BINARY_DIR to make macadam finds it

Signed-off-by: Luca Stocchi <lstocchi@redhat.com>