puller: Rework imageExtension()
This reuses as much as possible podman’s VMType.ImageFormat() so that we
automatically match podman expectations instead of duplicating this in
macadam.
The exceptions are WSL, where podman returns "tar" when we want
"tar.gz", qemu where we should be able to use both raw and qcow2
images and hyperv where we should be able to use both vhd and vhdx formats.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>